upgrading wiki

+++ profileinfo.php
@@ -28,8 +28,12 @@
 ini_set( 'zlib.output_compression', 'off' );
 
 $wgEnableProfileInfo = $wgProfileToDatabase = false;
+if ( isset( $_SERVER['MW_COMPILED'] ) ) {
+	require ( 'phase3/includes/WebStart.php' );
+} else {
+	require ( dirname( __FILE__ ) . '/includes/WebStart.php' );
+}
 
-require_once( './includes/WebStart.php' );
 
 header( 'Content-Type: text/html; charset=utf-8' );
 
@@ -70,6 +74,17 @@ header( 'Content-Type: text/html; charset=utf-8' );
 
 if ( !$wgEnableProfileInfo ) {
 	echo "<p>Disabled</p>\n";
+	echo "</body></html>";
+	exit( 1 );
+}
+
+$dbr = wfGetDB( DB_SLAVE );
+
+if( !$dbr->tableExists( 'profileinfo' ) ) {
+	echo "<p>No 'profileinfo' table exists, so we can't show you anything.</p>\n";
+	echo "<p>If you want to log profiling data, create the table using "
+		. "<tt>maintenance/archives/patch-profiling.sql</tt> and enable "
+		. "<tt>\$wgProfileToDatabase</tt>.</p>\n";
 	echo "</body></html>";
 	exit( 1 );
 }
@@ -214,8 +229,6 @@ $sort = 'time';
 if ( isset( $_REQUEST['sort'] ) && in_array( $_REQUEST['sort'], $sorts ) )
 	$sort = $_REQUEST['sort'];
 
-
-$dbr = wfGetDB( DB_SLAVE );
 $res = $dbr->select( 'profiling', '*', array(), 'profileinfo.php', array( 'ORDER BY' => 'pf_name ASC' ) );
 
 if (isset( $_REQUEST['filter'] ) )
